backend/requirements.txt
flask
flask-sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
flask-jwt-extended
flask-migrate
psycopg2-binary
python-dotenv
joblib
pandas
numpy
tensorflow
pillow
xgboost
gunicorn
flask-cors
requests
prometheus-client
scipy
scikit-learn
matplotlib
seaborn
paho-mqtt
redis




